% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.R
\name{binary}
\alias{binary}
\title{Convert Raster to Binary Map}
\usage{
binary(x, th)
}
\arguments{
\item{x}{A RasterLayer or SpatRaster object.}

\item{th}{A numeric threshold value between 0 and 1.}
}
\value{
A binary RasterLayer or SpatRaster where values greater than \code{th} are \code{TRUE}, and others are \code{FALSE}.
}
\description{
Converts a continuous raster dataset into a binary map based on a specified threshold.
}
\examples{
library(terra)
r <- rast(nrows=10, ncols=10, vals=runif(100))
binary_map <- binary(r, th=0.5)
plot(binary_map)
}
