% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Load Essential Libraries for habitat}
\usage{
.onAttach(libname, pkgname)
}
\description{
This file ensures that all necessary libraries are installed and loaded when the habitat package is loaded.
}
