% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_csv}
\alias{export_csv}
\title{Export Habitat Results to CSV}
\usage{
export_csv(result, file_path)
}
\arguments{
\item{result}{A list containing the habitat change metrics.}

\item{file_path}{The file path where the CSV file will be saved.}
}
\description{
Exports the habitat change metrics to a CSV file.
}
\examples{
export_csv(result, "habitat_results.csv")
}
