% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_txt}
\alias{export_txt}
\title{Export Habitat Results to TXT}
\usage{
export_txt(result, file_path)
}
\arguments{
\item{result}{A list containing the habitat change metrics.}

\item{file_path}{The file path where the TXT file will be saved.}
}
\description{
Exports the habitat change metrics to a TXT file.
}
\examples{
export_txt(result, "habitat_results.txt")
}
