% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat-package.R
\docType{package}
\name{habitat-package}
\alias{habitat}
\alias{habitat-package}
\title{habitat: An R Package for Analyzing and Comparing Habitat Changes}
\description{
An R package for analyzing and comparing habitat and species distributions changes. Ideal for students, ecologists, environmental researchers, and conservation planners.
}
\author{
\strong{Maintainer}: Saman Ghasemian Sorboni \email{samangh.edu@gmail.com} (\href{https://orcid.org/0000-0002-1184-1484}{ORCID})

Authors:
\itemize{
  \item Mehrdad Hadipour (\href{https://orcid.org/0000-0001-6026-8306}{ORCID})
}

}
\keyword{internal}
