% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidim.R
\name{hb_agg_md_raster}
\alias{hb_agg_md_raster}
\title{Aggregate Multidimensional Raster}
\usage{
hb_agg_md_raster(multidimensional_raster, fun = mean)
}
\arguments{
\item{multidimensional_raster}{A \code{SpatRaster} object representing the multidimensional dataset.}

\item{fun}{A function to apply for aggregation (e.g., mean, sum).}
}
\value{
A \code{SpatRaster} object representing the aggregated raster.
}
\description{
Summarizes the multidimensional dataset across a specific dimension (e.g., calculating the average habitat suitability over time).
This is the second step in a sequence of functions to analyze habitat trends.
}
\examples{
\dontrun{
# Same synthetic setup as in multidimensional
raster_files <- list(r1, r2)
time_points <- c(2022, 2023)

multidimensional_raster <- hb_multiD_raster(raster_files, time_points)
aggregated_raster <- hb_agg_md_raster(multidimensional_raster, fun = mean)
trends <- hb_analyze_changes(multidimensional_raster, fun = trend_analysis)

hb_plot_timesrs(trends, output_path = tempfile(fileext = ".tif"))
}
}
