% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_analyze.R
\name{hb_an_habitat}
\alias{hb_an_habitat}
\title{Analyze Suitable Habitat}
\usage{
hb_an_habitat(x1, x2, threshold = 0.5)
}
\arguments{
\item{x1}{Path to the input raster file (e.g., Elevation, NDVI).}

\item{x2}{Path to the habitat suitability raster file (e.g., ensemble map).}

\item{threshold}{Numeric value for the presence probability threshold. Default is 0.5.}
}
\value{
A list containing the range statistics and the histogram plot.
}
\description{
Analyzes the range of suitable habitat based on a given threshold and provides statistics and a histogram plot.
}
\details{
Calculates the range of suitable habitat areas using pre-aligned raster files.
}
\examples{
# Example usage
x1 <- "path/to/DEM.tif"
x2 <- "path/to/habitat.tif"
result <- hb_an_habitat(x1, x2, threshold = 0.5)
print(result$statistics)
print(result$plot)
}
\references{
Ghasemian Sorboni, S., Hadipour, M., & Pourebrahim, S (2024). habitat: An R Package for Analyzing and Comparing Habitat Changes. dataset. doi:.
}
