% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidim.R
\name{hb_analyze_changes}
\alias{hb_analyze_changes}
\title{Analyze Changes Using Statistical Summaries}
\usage{
hb_analyze_changes(multidimensional_raster, fun = trend)
}
\arguments{
\item{multidimensional_raster}{A \code{SpatRaster} object representing the multidimensional dataset.}

\item{fun}{A function to compute the statistical summary (e.g., trend, anomaly detection).}
}
\value{
A data frame containing the statistical summaries.
}
\description{
Applies statistical functions to identify trends, anomalies, or patterns in habitat changes.
This is the third step in a sequence of functions to analyze habitat trends.
}
\examples{
\dontrun{
raster_files <- list(r1, r2)
time_points <- c(2022, 2023)

multidimensional_raster <- hb_multiD_raster(raster_files, time_points)
aggregated_raster <- hb_agg_md_raster(multidimensional_raster, fun = mean)

# NOTE: Depending on your dataset, calculations may take time
trends <- hb_analyze_changes(multidimensional_raster, fun = trend_analysis)

hb_plot_timesrs(trends, output_path = tempfile(fileext = ".tif"))
}
}
