% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.R
\name{hb_binary}
\alias{hb_binary}
\title{Convert Raster to Binary Map}
\usage{
hb_binary(x, th)
}
\arguments{
\item{x}{A RasterLayer or SpatRaster object representing the continuous raster dataset to be converted.}

\item{th}{A numeric threshold value between 0 and 1, determining the cutoff point for converting values to \code{TRUE} or \code{FALSE}.}
}
\value{
A binary RasterLayer or SpatRaster where values greater than the specified threshold are \code{TRUE}, and values less than or equal to the threshold are \code{FALSE}.
}
\description{
Converts a continuous raster dataset into a binary map based on a specified threshold. Values in the raster that are greater than the threshold are converted to \code{TRUE}, while values less than or equal to the threshold are converted to \code{FALSE}.
}
\details{
This function is designed to convert a continuous raster dataset into a binary map. This is particularly useful for applications requiring binary classification of data, such as presence/absence mapping or suitability analysis.
}
\examples{
\dontrun{
# Example usage with a SpatRaster object

# Create a sample raster dataset with random values
r <- terra::rast(nrows = 10, ncols = 10, vals = runif(100))

# Convert the raster to a binary map using a threshold of 0.5
binary_map <- hb_binary(r, th = 0.5)

# Plot the resulting binary map
plot(binary_map, main = "Binary Map (Threshold = 0.5)")
}
}
