% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_area.R
\name{hb_calculate_area}
\alias{hb_calculate_area}
\title{Calculate Suitable and Unsuitable Areas}
\usage{
hb_calculate_area(binary_raster)
}
\arguments{
\item{binary_raster}{A SpatRaster object representing the binary raster with suitable (1/TRUE) and unsuitable (0/FALSE) areas.}
}
\value{
None. The results are printed in the console.
}
\description{
Calculates the suitable and unsuitable areas in hectares, square kilometers, and square meters from a binary raster.
}
\details{
This function calculates the total area of suitable and unsuitable regions in a binary raster and prints the results in hectares, square kilometers, and square meters.
}
\examples{
# Example usage with a binary SpatRaster object


# Create a sample binary raster
binary_raster <- rast(nrows = 10, ncols = 10, vals = sample(c(0, 1), 100, replace = TRUE))

# Calculate and print the suitable and unsuitable areas
hb_calculate_area(binary_raster)
}
