% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{hb_exp_txt}
\alias{hb_exp_txt}
\title{Export Habitat Results to TXT}
\usage{
hb_exp_txt(result, file_path)
}
\arguments{
\item{result}{A list containing the habitat change metrics. This list should include a component named \code{Compt.By.Models}, which holds the data to be exported.}

\item{file_path}{The file path where the TXT file will be saved. The path should include the desired file name and the \code{.txt} extension.}
}
\value{
None. This function is used for its side effect of writing the data to a TXT file.
}
\description{
Exports habitat change metrics to a TXT file. The habitat change metrics are extracted from the provided result list and saved as a TXT file at the specified file path.
}
\details{
Designed to facilitate the export of habitat change metrics to a TXT file, making it easier to save analysis results in a text format that can be readily shared and imported into other software for further analysis.
}
\examples{
# Example usage

# Create a sample result list with habitat change metrics
result <- list(Compt.By.Models = data.frame(Model = c("Model1", "Model2"),
Metric1 = c(0.1, 0.2), Metric2 = c(0.3, 0.4)))

# Export the habitat change metrics to a TXT file
hb_exp_txt(result, "habitat_results.txt")
}
