% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{hb_ext_eval}
\alias{hb_ext_eval}
\title{Extract Evaluation Metrics from sdmModels Object}
\usage{
hb_ext_eval(sdm_model)
}
\arguments{
\item{sdm_model}{An sdmModels object generated by the sdm function.}
}
\value{
A data frame with evaluation metrics.
}
\description{
Extracts evaluation metrics including threshold, sensitivity, specificity, TSS, MCC, F1, Kappa, NMI, PHI, PPV, NPV, CCR, MCR, Omission, Commission, and Prevalence from a trained sdm model.
}
\examples{
# Assume that the user has already trained their model and wants to extract the
threshold value based on metrics such as sensitivity (sp), specificity (se), max(sp + se),
 kappa, positive predictive value (ppv), negative predictive value (npv),
 normalized mutual information (nmi), correct classification rate (ccr), prevalence,
 and specific percentages (P10, P5, P1, P0).

 library(sdm)
# sdm_model <- trained_model_object

# Extract evaluation metrics
eval_metrics <- hb_ext_eval(sdm_model)
print(eval_metrics)
}
\references{
Ghasemian Sorboni, S., Hadipour, M., & Pourebrahim, S (2024). habitat: An R Package for Analyzing and Comparing Habitat Changes. dataset. doi:.
}
