% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{hb_merge}
\alias{hb_merge}
\title{Merge Rasters}
\usage{
hb_merge(raster_list, output_path = NULL)
}
\arguments{
\item{raster_list}{A list of \code{RasterLayer} objects to be merged.}

\item{output_path}{A character string representing the file path to save the merged raster. If \code{NULL}, the raster will not be saved to a file.}
}
\value{
A \code{RasterLayer} object representing the merged raster.
}
\description{
Merges two or more rasters into a single raster file. If you encounter any issues with \code{RasterLayer} objects, use the \code{spat_to_raster()} function to convert them.
}
\examples{
# Load sample rasters
raster1 <- raster("path/to/raster1.tif")
raster2 <- raster("path/to/raster2.tif")

# Merge rasters (two or more), and save it in '.tif' format (optional).
merged_raster <- hb_merge(list(raster1, raster2), "path/to/merged_raster.tif")

# NOTE: 'RasterLayer' issues? Use the spat_to_raster() function
raster1 <- spat_to_raster(rast("path/to/raster1.tif"))
raster2 <- spat_to_raster(rast("path/to/raster2.tif"))

# Plot the merged raster
plot(merged_raster)

}
