% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{hb_opt_th}
\alias{hb_opt_th}
\title{Calculate Optimal Thresholds from Evaluation Metrics}
\usage{
hb_opt_th(eval_metrics)
}
\arguments{
\item{eval_metrics}{A data frame with evaluation metrics.}
}
\value{
A list with optimal thresholds for different criteria.
}
\description{
Calculates various optimal thresholds based on different criteria such as max(sp + se), max(kappa), and others.
}
\examples{
# Assuming the user has already extracted evaluation metrics
PLEASE READ hb_ext_eval FIRST
# eval_metrics <- hb_ext_eval(sdm_model)

# Calculate optimal thresholds
thresholds <- hb_opt_th(eval_metrics)
print(thresholds)

# Example usage of the extracted threshold with existing functions:
# max_sp_se_threshold <- thresholds$max_sp_se
# binary_map <- hb_binary(r, th = max_sp_se_threshold)
# habitat_analysis <- hb_an_habitat(x1, x2, threshold = max_sp_se_threshold)
# range_analysis <- hb_range(x, y, th = max_sp_se_threshold)
}
\references{
Ghasemian Sorboni, S., Hadipour, M., & Pourebrahim, S (2024). habitat: An R Package for Analyzing and Comparing Habitat Changes. dataset. doi:.
}
