% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_habitat.R
\name{hb_plot}
\alias{hb_plot}
\title{Plot Habitat Map}
\usage{
hb_plot(
  raster,
  main = "Habitat Map",
  lonlat = TRUE,
  add_north_arrow = FALSE,
  background_color = "white",
  habitat_palette = "viridis",
  add_legend = TRUE
)
}
\arguments{
\item{raster}{A SpatRaster object representing the habitat map.}

\item{main}{Title for the plot.}

\item{lonlat}{Logical. If TRUE, plots the habitat map on a longitude and latitude scale.}

\item{add_north_arrow}{Logical. If TRUE, adds a North arrow to the plot.}

\item{background_color}{Character string specifying the background color of the map. Default is "white".}

\item{habitat_palette}{Character string specifying the palette for habitat areas.}

\item{add_legend}{Logical. If TRUE, includes a legend in the plot.}
}
\description{
Plots a habitat map with enhanced visualization options using ggplot2.
}
\details{
This function plots a habitat map with various optional parameters for enhanced visualization.
}
\examples{
# Example usage with a SpatRaster object

# Create a sample raster
raster <- rast(nrows = 10, ncols = 10, vals = runif(100))

# Plot habitat map
hb_plot(raster, main = "Sample Habitat Map", lonlat = TRUE, add_north_arrow = TRUE, background_color = "lightblue", add_legend = TRUE, habitat_palette = "viridis")
}
\references{
Ghasemian Sorboni, S., Hadipour, M., & Pourebrahim, S (2024). habitat: An R Package for Analyzing and Comparing Habitat Changes. dataset. doi:.
}
