% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2p.R
\name{hb_ras_to_pol}
\alias{hb_ras_to_pol}
\title{Convert Raster to Polygon}
\usage{
hb_ras_to_pol(raster, binary = FALSE)
}
\arguments{
\item{raster}{A SpatRaster object to be converted.}

\item{binary}{Logical. If \code{TRUE}, the raster will be treated as binary. Default is \code{FALSE}.}
}
\value{
An sf object representing the raster converted to polygons.
}
\description{
Converts a binary or continuous SpatRaster object to a polygon.
}
\details{
The function is designed to convert a raster dataset into a polygon, which is useful for visualizing raster data as vector data and performing vector-based spatial analyses.
}
\examples{
\dontrun{
# Create a sample binary raster or load it: "binary_raster"

# Convert the binary raster to polygons
binary_polygons <- hb_raster_to_polygon(binary_raster, binary = TRUE)
plot(binary_polygons)

# Create a sample raster or load it: "raster1"

# Convert the raster to polygons
continuous_polygons <- hb_ras_to_pol(raster1)
plot(continuous_polygons)
}
}
