% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2p.R
\name{hb_raster_to_polygon}
\alias{hb_raster_to_polygon}
\title{Convert Raster to Polygon}
\usage{
hb_raster_to_polygon(raster, binary = FALSE)
}
\arguments{
\item{raster}{A SpatRaster object to be converted.}

\item{binary}{Logical. If \code{TRUE}, the raster will be treated as binary. Default is \code{FALSE}.}
}
\value{
An sf object representing the raster converted to polygons.
}
\description{
Converts a binary or continuous SpatRaster object to a polygon.
}
\details{
The function is designed to convert a raster dataset into a polygon, which is useful for visualizing raster data as vector data and performing vector-based spatial analyses.
}
\examples{
# Example usage with SpatRaster objects

# Create a sample binary raster
binary_raster <- rast(nrows = 10, ncols = 10, vals = sample(c(0, 1), 100, replace = TRUE))

# Convert the binary raster to polygons
binary_polygons <- hb_raster_to_polygon(binary_raster, binary = TRUE)

# Plot the resulting polygons
plot(binary_polygons)

# Create a sample continuous raster
continuous_raster <- rast(nrows = 10, ncols = 10, vals = runif(100))

# Convert the continuous raster to polygons
continuous_polygons <- hb_raster_to_polygon(continuous_raster)

# Plot the resulting polygons
plot(continuous_polygons)
}
