% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.R
\name{hb_sstat}
\alias{hb_sstat}
\title{Summary Statistics for Raster Data}
\usage{
hb_sstat(raster)
}
\arguments{
\item{raster}{A SpatRaster object representing the raster dataset for which summary statistics will be generated.}
}
\value{
A list containing summary statistics of the raster values, including the mean, median, standard deviation, minimum, and maximum of the raster values.
}
\description{
Provides summary statistics for a given SpatRaster object, calculating key statistical measures to summarize the distribution of values within the raster dataset.
}
\details{
Designed to take a raster dataset and compute summary statistics that provide insights into the data's distribution and variability. These statistics are useful for understanding the overall characteristics of the raster data.
}
\examples{
\dontrun{
# Create a sample raster dataset or load it: "raster1"

# Generate the summary statistics for the raster
raster_summary <- hb_sstat(raster1)

# Display the summary statistics
print(raster_summary)
}
}
