% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{hb_stack_dfs}
\alias{hb_stack_dfs}
\title{Stack Data Frames}
\usage{
hb_stack_dfs(dfs)
}
\arguments{
\item{dfs}{A list of data frames to be stacked.}
}
\value{
A single data frame that combines all input data frames by row binding.
}
\description{
Stacks multiple data frames into a single data frame.
}
\details{
This function takes a list of data frames and stacks them into a single data frame by row binding. It is useful for combining multiple data frames for further analysis.
}
\examples{
\dontrun{
# Create sample data frames or load it: "df1" and "df2"

# Stack the data frames
stacked_df <- hb_stack_dfs(list(df1, df2))
print(stacked_df)
}
}
