% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_range.R
\name{HC_plot}
\alias{HC_plot}
\title{Plot Habitat Changes}
\usage{
HC_plot(data)
}
\arguments{
\item{data}{A data frame containing the habitat change metrics.}
}
\description{
Plots the habitat changes as a bar chart.
}
\examples{
HC_plot(result$Compt.By.Models)
}
