% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changes.R
\name{raster_to_spat}
\alias{raster_to_spat}
\title{Convert RasterLayer to SpatRaster}
\usage{
raster_to_spat(raster_layer)
}
\arguments{
\item{raster_layer}{A RasterLayer object.}
}
\value{
A SpatRaster object.
}
\description{
Converts a RasterLayer object to a SpatRaster object.
}
