% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_check.R
\name{res_check}
\alias{res_check}
\title{Validate Raster Inputs}
\usage{
res_check(x, y)
}
\arguments{
\item{x}{A RasterLayer or SpatRaster object representing the first dataset.}

\item{y}{A RasterLayer or SpatRaster object representing the second dataset.}
}
\value{
Returns \code{TRUE} if all checks pass, otherwise throws an error.
}
\description{
Checks whether two raster datasets have identical extent, CRS, dimensions, and resolution.
}
\examples{
library(terra)
r1 <- rast(nrows=10, ncols=10, vals=runif(100))
r2 <- rast(nrows=10, ncols=10, vals=runif(100))
res_check(r1, r2) # Should return TRUE
}
