% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changes.R
\name{spat_to_raster}
\alias{spat_to_raster}
\title{Convert SpatRaster to RasterLayer}
\usage{
spat_to_raster(spat_raster)
}
\arguments{
\item{spat_raster}{A SpatRaster object.}
}
\value{
A RasterLayer object.
}
\description{
Converts a SpatRaster object to a RasterLayer object.
}
