% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{hb_exp_raster}
\alias{hb_exp_raster}
\title{Export Raster to File}
\usage{
hb_exp_raster(raster_data, file_path)
}
\arguments{
\item{raster_data}{A raster (SpatRaster) object to be exported.}

\item{file_path}{The file path where the raster will be saved. The extension should be .tif, .png, or .jpg.}
}
\value{
None. This function is used for its side effect of writing the raster to a file.
}
\description{
Exports a raster to a specified file format (.tif, .png, or .jpg).
}
\examples{
# Load sample raster data
raster_data <- rast("path/to/raster.tif")

# Export the raster to a .tif file
hb_exp_raster(raster_data, "output_raster.tif")

# Export the raster to a .png file
hb_exp_raster(raster_data, "output_raster.png")

# Export the raster to a .jpg file
hb_exp_raster(raster_data, "output_raster.jpg")
}
