% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{hb_merge}
\alias{hb_merge}
\title{Merge Rasters}
\usage{
hb_merge(raster_list, output_path = NULL)
}
\arguments{
\item{raster_list}{A list of \code{RasterLayer} objects to be merged.}

\item{output_path}{A character string representing the file path to save the merged raster. If \code{NULL}, the raster will not be saved to a file.}
}
\value{
A \code{RasterLayer} object representing the merged raster.
}
\description{
Merges two or more rasters into a single raster file. If you encounter any issues with \code{RasterLayer} objects, use the \code{spat_to_raster()} function to convert them.
}
\examples{
\dontrun{
# Create two sample rasters or load it: "raster1" and "raster2"
# Merge rasters (two or more), optionally saving to a file
merged_raster <- hb_merge(list(raster1, raster2),
                           tempfile(fileext = ".tif"))

# NOTE: 'RasterLayer' issues? Use the spat_to_raster() function
raster1_rl <- spat_to_raster(raster1)
raster2_rl <- spat_to_raster(raster2)

# Plot the merged raster
plot(merged_raster)
}
}
