% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changes.R
\name{hb_plot_changes}
\alias{hb_plot_changes}
\title{Plot Habitat Changes (terra)}
\usage{
hb_plot_changes(
  changes_raster,
  title = "Changes",
  xlab = "Longitude",
  ylab = "Latitude",
  bg_color = "white"
)
}
\arguments{
\item{changes_raster}{A SpatRaster object generated by \code{hb_changes}.}

\item{title}{Optional. Title of the plot. Default is "Changes".}

\item{xlab}{Optional. Label for the x-axis. Default is "Longitude".}

\item{ylab}{Optional. Label for the y-axis. Default is "Latitude".}

\item{bg_color}{Optional. Background color of the plot region and NA cells. Default is "white".}
}
\description{
Plots a SpatRaster highlighting areas of habitat gain, loss, stability, and absence using terra's native plotting.
}
\details{
Uses terra::plot() for fast, seam-free visualization of rasters.
}
\examples{
# changes_raster <- hb_changes(raster1_binary, raster2_binary)
# hb_plot_changes(changes_raster, title = "Habitat Changes")
}
