% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.R
\name{hb_sstat}
\alias{hb_sstat}
\title{Summary Statistics for Raster Data}
\usage{
hb_sstat(raster)
}
\arguments{
\item{raster}{A SpatRaster object representing the raster dataset for which summary statistics will be generated.}
}
\value{
A list containing summary statistics of the raster values, including the mean, median, standard deviation, minimum, and maximum of the raster values.
}
\description{
Provides summary statistics for a given SpatRaster object, calculating key statistical measures to summarize the distribution of values within the raster dataset.
}
\details{
Designed to take a raster dataset and compute summary statistics that provide insights into the data's distribution and variability. These statistics are useful for understanding the overall characteristics of the raster data.
}
\examples{
# Example usage with a SpatRaster object

# Create a sample raster dataset with random values
raster <- rast(nrows = 10, ncols = 10, vals = runif(100))

# Generate the summary statistics for the raster
raster_summary <- hb_sstat(raster)

# Display the summary statistics
print(raster_summary)
}
\references{
Ghasemian Sorboni, S., Hadipour, M., & Pourebrahim, S (2024). habitat: An R Package for Analyzing and Comparing Habitat Changes. dataset. doi:.
}
