% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{hb_stack_rasters}
\alias{hb_stack_rasters}
\title{Stack Raster Layers}
\usage{
hb_stack_rasters(rasters)
}
\arguments{
\item{rasters}{A list of SpatRaster objects to be stacked.}
}
\value{
A SpatRaster object that combines all input raster layers.
}
\description{
Stacks multiple SpatRaster objects into a single SpatRaster object.
}
\details{
This function takes a list of SpatRaster objects and stacks them into a single SpatRaster object. It is useful for combining multiple raster layers for further analysis.
}
\examples{
# Example usage with SpatRaster objects

# Create sample raster datasets
raster1 <- rast(nrows = 10, ncols = 10, vals = runif(100))
raster2 <- rast(nrows = 10, ncols = 10, vals = runif(100))

# Stack the rasters
stacked_rasters <- hb_stack_rasters(list(raster1, raster2))
plot(stacked_rasters)
}
